+++ CHANGES.md
@@ -1,7 +1,7 @@
 # Changes
 
 ## Unreleased
-- Added `Dataset::gcps` and `Dataset::set_gcps` APIs
+- Added `Dataset::gcp_projection`, `Dataset::gcps`, `Dataset::set_gcps` APIs
 
   - <https://github.com/georust/gdal/pull/404>